worker/requirements.txt
redis>=4.5.1
torch>=2.0.0
diffusers>=0.21.4
transformers>=4.25.1
accelerate>=0.16.0
safetensors>=0.3.1
pillow>=9.4.0
numpy>=1.24.2
boto3>=1.28.0
python-dotenv>=1.0.0